+++ frontend/src/components/Layout/index.tsx
@@ -1,5 +1,7 @@
-import Flex from '../Flex';
-import Input from '../Input';
+import Flex from '../common/Flex';
+import Input from '../common/Input';
+import Space from '../common/Space';
+import Logo from './Logo';
 
 type LayoutProps = {
   children: React.ReactNode;
@@ -14,6 +16,8 @@ const Layout = ({ children }: LayoutProps) => {
       $backgroundColor="white"
       padding={4}
     >
+      <Logo />
+      <Space size={5} />
       <Input />
       <Flex $flexDirection="column" overflow="scroll">
         {children}
