+++ frontend/src/pages/RootPage.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Layout from '../components/common/Layout';
+import Layout from '../components/Layout';
 
 const RootPage = () => {
   return (
